refactor: extract CLI utils for easier testing
