+++ coverage.txt
@@ -3,7 +3,7 @@
 ℹ file           | line % | branch % | funcs % | uncovered lines
 ℹ ---------------------------------------------------------------
 ℹ src            |        |          |         | 
-ℹ  cli.js        |  98.39 |    96.15 |  100.00 | 123-124
+ℹ  cli-utils.js  | 100.00 |   100.00 |  100.00 | 
 ℹ  config.js     | 100.00 |    89.19 |  100.00 | 
 ℹ  constants.js  | 100.00 |   100.00 |  100.00 | 
 ℹ  fileserver.js | 100.00 |    90.91 |  100.00 | 
@@ -13,6 +13,6 @@
 ℹ  response.js   | 100.00 |    98.44 |  100.00 | 
 ℹ  woodland.js   | 100.00 |    93.26 |  100.00 | 
 ℹ ---------------------------------------------------------------
-ℹ all files      |  99.93 |    96.17 |   99.32 | 
+ℹ all files      | 100.00 |    96.28 |   99.31 | 
 ℹ ---------------------------------------------------------------
 ℹ end of coverage report
